+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/MultiSetCompanionTest.kt
@@ -1,6 +1,6 @@
 package xyz.lbres.kotlinutils.set.multiset
 
-import xyz.lbres.kotlinutils.set.multiset.testimpl.TestMultiSet
+import xyz.lbres.kotlinutils.set.multiset.testutils.TestMultiSet
 import kotlin.test.Test
 import kotlin.test.assertEquals
