+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/constimmutable/runConstBinaryTests.kt
@@ -1,111 +1,38 @@
-package xyz.lbres.kotlinutils.set.multiset.constimpl.immutable
+package xyz.lbres.kotlinutils.set.multiset.constimmutable
 
-import xyz.lbres.kotlinutils.list.IntList
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
-import xyz.lbres.kotlinutils.set.multiset.testimpl.TestMultiSet
-import xyz.lbres.kotlinutils.set.multiset.testimpl.TestMutableMultiSet
+import xyz.lbres.kotlinutils.set.multiset.testutils.TestMultiSet
 import kotlin.test.assertEquals
-import kotlin.test.assertFalse
-import kotlin.test.assertNotEquals
 
 private val e1 = ArithmeticException()
 private val e2 = NullPointerException()
 private val e3 = IllegalArgumentException()
 
-fun runConstEqualsTests() {
-    // equals
-    var set1: ConstMultiSet<Int> = constMultiSetOf()
-    assertEquals(set1, set1)
-
-    set1 = constMultiSetOf(3)
-    assertEquals(set1, set1)
-
-    set1 = constMultiSetOf(3, 3, 3)
-    assertEquals(set1, set1)
-
-    set1 = constMultiSetOf(2, 3, 4)
-    assertEquals(set1, set1)
-
-    set1 = constMultiSetOf(1, 2, 3)
-    var set2 = multiSetOf(3, 1, 2)
-    assertEquals(set1, set2)
-    assertEquals(set2, set1)
-
-    var otherSet = TestMultiSet(listOf(1, 2, 3))
-    assertEquals<MultiSet<Int>>(set1, otherSet)
-
-    // not equals
-    set1 = constMultiSetOf()
-    set2 = constMultiSetOf(0)
-    assertNotEquals(set1, set2)
-    assertNotEquals(set2, set1)
-
-    set1 = constMultiSetOf(1, 1)
-    set2 = constMultiSetOf(1)
-    assertNotEquals(set1, set2)
-    assertNotEquals(set2, set1)
-
-    set1 = constMultiSetOf(1, 2)
-    set2 = constMultiSetOf(2, 2)
-    assertNotEquals(set1, set2)
-    assertNotEquals(set2, set1)
-
-    set1 = constMultiSetOf(-1, 3, 1, -3)
-    otherSet = TestMultiSet(listOf(2, -2))
-    assertNotEquals<MultiSet<Int>>(set1, otherSet)
-
-    // other type
-    val stringSet1 = constMultiSetOf("", "abc")
-    assertEquals(stringSet1, stringSet1)
-
-    val stringSet2 = constMultiSetOf("abc")
-    assertNotEquals(stringSet1, stringSet2)
-    assertNotEquals(stringSet2, stringSet1)
-
-    val listSet1 = constMultiSetOf(listOf(12, 34, 56), listOf(77, 78, 0, 15), listOf(5))
-    assertEquals(listSet1, listSet1)
-
-    val listSet2 = constMultiSetOf(listOf(12, 34, 56))
-    assertNotEquals(listSet1, listSet2)
-    assertNotEquals(listSet2, listSet1)
-
-    assertFalse(stringSet1 == set1)
-
-    val otherListSet = TestMultiSet(listOf(listOf(12)))
-    assertNotEquals<MultiSet<IntList>>(listSet2, otherListSet)
-
-    // mutable
-    set1 = constMultiSetOf(1, 2, 3)
-    var mutableSet = mutableMultiSetOf(1, 2, 3)
-    assertEquals<MultiSet<Int>>(set1, mutableSet)
-
-    mutableSet = mutableMultiSetOf(1)
-    assertNotEquals<MultiSet<Int>>(set1, mutableSet)
-
-    val otherMutableSet = TestMutableMultiSet(listOf(1, 2, 3))
-    assertEquals<MultiSet<Int>>(set1, otherMutableSet)
-}
-
 fun runConstMinusTests() {
     // empty
-    var intSet1 = constMultiSetOf<Int>()
-    var intSet2 = constMultiSetOf<Int>()
-    assertEquals(constMultiSetOf(), intSet1 - intSet2)
-    assertEquals(constMultiSetOf(), intSet2 - intSet1)
+    var intSet1: ConstMultiSet<Int> = emptyConstMultiSet()
+    var intSet2: ConstMultiSet<Int> = emptyConstMultiSet()
+    assertEquals(emptyConstMultiSet(), intSet1 - intSet2)
+    assertEquals(emptyConstMultiSet(), intSet2 - intSet1)
 
     intSet1 = constMultiSetOf(1, 2, 3, 3)
-    assertEquals(intSet1, intSet1 - intSet2)
-    assertEquals(constMultiSetOf(), intSet2 - intSet1)
+    assertEquals(intSet1, intSet1 - emptyConstMultiSet())
+    assertEquals(emptyConstMultiSet(), emptyConstMultiSet<Int>() - intSet1)
 
     // equal
     intSet1 = constMultiSetOf(1, 2, 3, 4, 5)
-    assertEquals(constMultiSetOf(), intSet1 - intSet1)
+    assertEquals(emptyConstMultiSet(), intSet1 - intSet1)
+
+    intSet1 = constMultiSetOf(1, 2, 3)
+    intSet2 = constMultiSetOf(3, 1, 2)
+    assertEquals(emptyConstMultiSet(), intSet1 - intSet2)
+    assertEquals(emptyConstMultiSet(), intSet2 - intSet1)
 
     var listSet1 = constMultiSetOf(listOf(1, 2, 3), listOf(456, 789))
-    assertEquals(constMultiSetOf(), listSet1 - listSet1)
+    assertEquals(emptyConstMultiSet(), listSet1 - listSet1)
 
     var otherSet = TestMultiSet(listOf(1, 2, 3, 4, 5))
-    assertEquals(constMultiSetOf(), intSet1 - otherSet)
+    assertEquals(emptyConstMultiSet(), intSet1 - otherSet)
 
     // all shared
     intSet1 = constMultiSetOf(1, 1, 2, 3, 4, 4, 4)
@@ -174,15 +101,15 @@ fun runConstMinusTests() {
 
 fun runConstPlusTests() {
     // empty
-    var intSet1 = constMultiSetOf<Int>()
-    var intSet2 = constMultiSetOf<Int>()
-    assertEquals(constMultiSetOf(), intSet1 + intSet2)
-    assertEquals(constMultiSetOf(), intSet2 + intSet1)
+    var intSet1: ConstMultiSet<Int> = emptyConstMultiSet()
+    var intSet2: ConstMultiSet<Int> = emptyConstMultiSet()
+    assertEquals(emptyConstMultiSet(), intSet1 + intSet2)
+    assertEquals(emptyConstMultiSet(), intSet2 + intSet1)
 
     // non-empty
     intSet1 = constMultiSetOf(1, 2, 3, 3)
-    assertEquals(intSet1, intSet1 + intSet2)
-    assertEquals(intSet1, intSet2 + intSet1)
+    assertEquals(intSet1, intSet1 + emptyConstMultiSet())
+    assertEquals(intSet1, emptyConstMultiSet<Int>() + intSet1)
 
     intSet1 = constMultiSetOf(1)
     intSet2 = constMultiSetOf(1)
@@ -226,28 +153,28 @@ fun runConstPlusTests() {
 
 fun runConstIntersectTests() {
     // empty
-    var intSet1 = constMultiSetOf<Int>()
+    var intSet1: ConstMultiSet<Int> = emptyConstMultiSet()
 
-    var intSet2 = constMultiSetOf<Int>()
-    assertEquals(constMultiSetOf(), intSet1 intersect intSet2)
+    var intSet2: ConstMultiSet<Int> = emptyConstMultiSet()
+    assertEquals(emptyConstMultiSet(), intSet1 intersect intSet2)
 
     intSet2 = constMultiSetOf(1, 2, 3)
-    assertEquals(constMultiSetOf(), intSet1 intersect intSet2)
-    assertEquals(constMultiSetOf(), intSet2 intersect intSet1)
+    assertEquals(emptyConstMultiSet(), intSet1 intersect intSet2)
+    assertEquals(emptyConstMultiSet(), intSet2 intersect intSet1)
 
     // none shared
     intSet1 = constMultiSetOf(1, 2, 3)
     intSet2 = constMultiSetOf(4, 5, 6, 7, 8)
-    assertEquals(constMultiSetOf(), intSet1 intersect intSet2)
-    assertEquals(constMultiSetOf(), intSet2 intersect intSet1)
+    assertEquals(emptyConstMultiSet(), intSet1 intersect intSet2)
+    assertEquals(emptyConstMultiSet(), intSet2 intersect intSet1)
 
     var otherSet = TestMultiSet(listOf(4, 5, 6, 7, 8))
-    assertEquals(constMultiSetOf(), intSet1 intersect otherSet)
+    assertEquals(emptyConstMultiSet(), intSet1 intersect otherSet)
 
     var listSet1 = constMultiSetOf(listOf(1, 2, 3), listOf(4, 5))
     var listSet2 = constMultiSetOf(listOf(1, 2), listOf(3, 4, 5))
-    assertEquals(constMultiSetOf(), listSet1 intersect listSet2)
-    assertEquals(constMultiSetOf(), listSet2 intersect listSet1)
+    assertEquals(emptyConstMultiSet(), listSet1 intersect listSet2)
+    assertEquals(emptyConstMultiSet(), listSet2 intersect listSet1)
 
     // all shared
     intSet1 = constMultiSetOf(1, 2, 3)
