+++ kotlin-utils/src/test/kotlin/xyz/lbres/kotlinutils/set/multiset/constmutable/runMutableContainsTests.kt
@@ -1,4 +1,4 @@
-package xyz.lbres.kotlinutils.set.multiset.constimpl.mutable
+package xyz.lbres.kotlinutils.set.multiset.constmutable
 
 import xyz.lbres.kotlinutils.set.multiset.* // ktlint-disable no-wildcard-imports no-unused-imports
 import kotlin.test.assertFalse
